Refactor: 게임에서 계산부를 다른 객체로 분리
